Migrated CSVN to a separate crate.
