+++ crates/tool-csvn/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "csvn"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "CSV Normalizer tool."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+string-interner = "0.19.0"
+anyhow = "1.0.99"
+clap = "4.5.45"
+chrono = "0.4.41"
+csv = "1.3.1"
+tracing = "0.1.41"
