+++ crates/tool-csvn/src/csvn_app.rs
@@ -1,4 +1,3 @@
-use crate::tools::csvn::models::CsvNConfig;
 use anyhow::{anyhow, Context, Result};
 use chrono::{DateTime, Utc};
 use csv::{StringRecord, Writer, WriterBuilder};
@@ -13,6 +12,7 @@ use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use string_interner::DefaultSymbol;
+use crate::models::CsvNConfig;
 
 /// Determines headers for CSV processing.
 ///
