+++ crates/tools/Cargo.toml
@@ -19,10 +19,6 @@ name = "ts"
 path = "src/bin/ts.rs"
 
 [[bin]]
-name = "csvn"
-path = "src/bin/csvn.rs"
-
-[[bin]]
 name = "get-lines"
 path = "src/bin/get_lines.rs"
