--- crates/tools/src/tools/csvn/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod csvn_app;
-pub mod models;
